refactor(front/app): Add search component
