+++ guifaffinity-front/src/components/header.tsx
